hud-service: delay dropping of query by 1 second
Unity sends a 'CloseQuery' followed immediately by a 'StartQuery' on
each keystroke in the search field.  This results in a bunch of extra
dbusmenu messages being sent as the use count drops briefly to 0 only to
go immediately back to 1.

We can prevent that by delaying the destruction of the query.

This decreases the number of dbus messages resulting from a short search
by a factor of about 4 (on a number that measures in the 100s).