hud-service: dbusmenu collector: hold a reference on the root menu item
In the root-changed case, the DbusmenuClient drops the reference on the
root item before alerting us that it has changed.  We then go iterating
over our own (now-dead) copy of that item in order to remove it from our
internal hashtable.

Keep our own reference to it to ensure that this is safe.

This could explain lp:955937.