Increase the default swap penalty to 15
The current value is too low.  It allows "xyze" as a valid match for "File" (3
swapped characters at a cost of 10 each gives a total cost of 30, which is the
maximum permitted value).

It also causes an issue with accents.  Consider the case of typing
"pref" and encountering the menu items "préférence" and "précédent".
Clearly the user wants "préférence".  During normalisation, we convert
each of these strings to "pre'fe'rence" and "pre'ce'dent".

"pref" vs "pre'fe'rence" is best scored as the first "'" being dropped
and the trailing "e'rence" also being dropped.  10 points for the accent
drop plus 7 for the trailing "e'rence".  Total of 17 points.

"pref" vs "pre'ce'dent" seems like it should be scored worse since the
accent is dropped and the following character is incorrect.  Instead,
however, the algorithm assigns a lower score for the possibility that
the accent is replaced by the letter "f" (10 points) and then the rest
of the characters "ce'dent" are dropped (7 points).  Also a total of 17
points.

This tie would be prevented with a higher swap cost.

This problem could be fixed by special-casing the dropping of accents in
the core search algorithm and I tested this approach but it had a
substantial performance cost on the rather tight inner loop (nearly
doubling the amount of time to do a search).

Another possibility would be to drop all accents entirely during the
normalisation process.  This sounds acceptable for French, but I imagine
other languages may find this to be less than ideal.

Also, consider "fl" against "file" and "fix" as another example that
doesn't rely on accents to make its case (although the argument is
somewhat weaker here).  This shows that we should probably just increase
the swap cost.